Processing Speed and Secret Sort Added
Secrets were updated to be sorted and processed from largest to smallest, this should avoid issues where a smaller secrets was contained in a larger secret and was being removed first causing the larger secret to not be removed.

Updated the processing code to increase overall speed.

Added tests for both processing.go and utils.go.

Updated GitHub Actions to remove Garble and include cheksums for the files.